Finished Project with main methods
In this project, I use many an int variable- score to recored how many times a player have scored. But in order to simplify comparing procedure, when 2 players both score 4 times, it can be view as they score 40-40, that is to say we can consider that they both just score 3 times.
I am a Java beginer, so I don´t know how to simplify those `if-else´sentences. I hope to discuss with you all!